+++ packages/frameworks/vue/src/date-picker/date-picker-table-context.ts
@@ -1,0 +1,7 @@
+import type { TableProps } from '@zag-js/date-picker'
+import { createContext } from '../context'
+
+export interface DatePickerTableContext extends TableProps {}
+
+export const [DatePickerTableProvider, useDatePickerTableContext] =
+  createContext<DatePickerTableContext>('DatePickerTableContext')