+++ packages/frameworks/vue/src/date-picker/date-picker-trigger.tsx
@@ -1,0 +1,18 @@
+import { defineComponent } from 'vue'
+import { ark, type HTMLArkProps } from '../factory'
+import { useDatePickerContext } from './date-picker-context'
+
+export interface DatePickerTriggerProps extends HTMLArkProps<'button'> {}
+
+export const DatePickerTrigger = defineComponent({
+  name: 'DatePickerTrigger',
+  setup(_, { attrs, slots }) {
+    const api = useDatePickerContext()
+
+    return () => (
+      <ark.button {...api.value.triggerProps} {...attrs}>
+        {slots.default?.()}
+      </ark.button>
+    )
+  },
+})